Fixed VS build.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@355 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/fem/fem_cards_beuslo.cpp
@@ -118,7 +118,7 @@ namespace dnvgl {
                            const long &SIDE,
                            const std::deque<double> &RLOAD,
                            const std::deque<double> &ILOAD) :
-               beuslo(LLC, LOTYP, COMPLX, LAYER, ELNO, RLOAD.size(),
+                           beuslo(LLC, LOTYP, COMPLX, LAYER, ELNO, static_cast<long>(RLOAD.size()),
                       INTNO, SIDE, RLOAD, ILOAD) {}
 
             beuslo::beuslo(const long &LLC,
@@ -141,7 +141,7 @@ namespace dnvgl {
                            const long &SIDE,
                            const std::deque<double> &RLOAD,
                            const std::deque<double> &ILOAD) :
-               beuslo(LLC, LOTYP, ILOAD.size() > 0, LAYER, ELNO, RLOAD.size(),
+               beuslo(LLC, LOTYP, ILOAD.size() > 0, LAYER, ELNO, static_cast<long>(RLOAD.size()),
                       INTNO, SIDE, RLOAD, ILOAD) {}
 
             const dnvgl::extfem::fem::cards::types
